refactor: removes the request forwarding logic for the configuration module
