+++ config/src/main/java/com/alibaba/nacos/config/server/configuration/NacosConfigConfiguration.java
@@ -16,7 +16,7 @@
 package com.alibaba.nacos.config.server.configuration;
 
 import com.alibaba.nacos.config.server.filter.NacosWebFilter;
-import com.alibaba.nacos.config.server.filter.TransferToLeaderFilter;
+import com.alibaba.nacos.config.server.filter.CurcuitFilter;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
@@ -49,18 +49,18 @@ public class NacosConfigConfiguration {
 	@Conditional(ConditionDistributedEmbedStorage.class)
 	@Bean
 	public FilterRegistrationBean transferToLeaderRegistration() {
-		FilterRegistrationBean<TransferToLeaderFilter> registration = new FilterRegistrationBean<>();
+		FilterRegistrationBean<CurcuitFilter> registration = new FilterRegistrationBean<>();
 		registration.setFilter(transferToLeader());
 		registration.addUrlPatterns("/v1/cs/*");
-		registration.setName("transferToLeader");
+		registration.setName("curcuitFilter");
 		registration.setOrder(6);
 		return registration;
 	}
 
 	@Conditional(ConditionDistributedEmbedStorage.class)
 	@Bean
-	public TransferToLeaderFilter transferToLeader() {
-		return new TransferToLeaderFilter();
+	public CurcuitFilter transferToLeader() {
+		return new CurcuitFilter();
 	}
 
 }
