+++ config/src/main/java/com/alibaba/nacos/config/server/model/event/RaftDBErrorRecoverEvent.java
@@ -21,7 +21,6 @@ import com.alibaba.nacos.core.notify.Event;
 
 /**
  *
- *
  * @author <a href="mailto:liaochuntao@live.com">liaochuntao</a>
  */
 @SuppressWarnings("PMD.ClassNamingShouldBeCamelRule")
